refactor(doc): Re-inline Node<T> + bracket parsing into StructTree
