+++ src/test/java/net/boyechko/pdf/autoa11y/fixes/WrapParagraphRunInListTest.java
@@ -24,7 +24,7 @@ import com.itextpdf.kernel.pdf.tagging.PdfStructElem;
 import java.util.List;
 import net.boyechko.pdf.autoa11y.PdfTestBase;
 import net.boyechko.pdf.autoa11y.document.DocContext;
-import net.boyechko.pdf.autoa11y.document.TreeDiagram;
+import net.boyechko.pdf.autoa11y.document.StructTree;
 import org.junit.jupiter.api.Test;
 
 class WrapParagraphRunInListTest extends PdfTestBase {
@@ -40,7 +40,7 @@ class WrapParagraphRunInListTest extends PdfTestBase {
                     document.addKid(p2);
                     document.addKid(p3);
                     assertEquals(
-                            "Document[P[], P[], P[]]", TreeDiagram.toRoleTree(document).toString());
+                            "Document[P[], P[], P[]]", StructTree.toRoleTree(document).toString());
 
                     DocContext ctx = new DocContext(pdfDoc);
                     WrapParagraphRunInList fix =
@@ -49,7 +49,7 @@ class WrapParagraphRunInListTest extends PdfTestBase {
 
                     assertEquals(
                             "Document[L[LI[LBody[P[]]], LI[LBody[P[]]], LI[LBody[P[]]]]]",
-                            TreeDiagram.toRoleTree(document).toString());
+                            StructTree.toRoleTree(document).toString());
                 });
     }
 
@@ -64,7 +64,7 @@ class WrapParagraphRunInListTest extends PdfTestBase {
                     document.addKid(p2);
                     document.addKid(p3);
                     assertEquals(
-                            "Document[P[], P[], P[]]", TreeDiagram.toRoleTree(document).toString());
+                            "Document[P[], P[], P[]]", StructTree.toRoleTree(document).toString());
 
                     DocContext ctx = new DocContext(pdfDoc);
                     WrapParagraphRunInList fix =
@@ -73,7 +73,7 @@ class WrapParagraphRunInListTest extends PdfTestBase {
 
                     assertEquals(
                             "Document[L[LI[LBody[P[]]], LI[LBody[P[]]], LI[LBody[P[]]]]]",
-                            TreeDiagram.toRoleTree(document).toString());
+                            StructTree.toRoleTree(document).toString());
                 });
     }
 }
